pyproject.toml
[project]
name = "aegion-poc-crewai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "crewai[memory,tools]>=0.203.1",
    "fastapi>=0.119.0",
    "openai>=1.109.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "slack-sdk>=3.37.0",
    "uvicorn>=0.37.0",
]
